% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pedRSP.R
\name{computeAELRTB}
\alias{computeAELRTB}
\title{Compute an approximation (AELRTB) to the expected likelihood ratio test (ELRT) and power from a pedRSP object.}
\usage{
computeAELRTB(pedRSP, truehsq, null, sig.level = 0.05)
}
\arguments{
\item{pedRSP}{a pedRSP object computed using the computeRSPs function}

\item{truehsq}{a vector of at least length 1 of the true heritability.  Exactly one of truehsq and null must have length 1.}

\item{null}{a vector of at least length 1 of the null hypothesis heritability.}

\item{sig.level}{a scalar of the significance level of the hypothesis test}
}
\value{
a pedELRT object with the approximated ELRT and power.
}
\description{
This function computes an approximation (AELRTB) to the expectation of the likelihood ratio test statistic from a pedRSP object and uses this to calculate power. This function uses the variance of the log-eigenvalues and geometric mean of eigenvalues to compute the ELRT.
}
\details{
This function calculates the approximate ELRT/power from a pedRSP object using only the actual study size and the variance of the log-eigenvalues.
This approximation can both over and under estimate the exact ELRT depending on the values of truehsq and null.  See paper for more details
}
\examples{
require(kinship2)
test1 <- data.frame(id  =c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14),
mom =c(0, 0, 0, 0, 2, 2, 4, 4, 6,  2,  0,  0, 12, 13),
dad =c(0, 0, 0, 0, 1, 1, 3, 3, 3,  7,  0,  0, 11, 10),
sex =c(0, 1, 0, 1, 0, 1, 0, 1, 0,  0,  0,  1,  1,  1))
tped <- with(test1, pedigree(id, dad, mom, sex))
x <- computeRSPs(ped=tped)
elrtb <- computeAELRTB(pedRSP=x,truehsq=0.3,null=seq(0.1,0.8,0.1))
plot(elrtb)
}
\author{
Jesse D. Raffa
}
\seealso{
\code{\link[pedRSP]{computeRSPs}}
}

