% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pedRSP.R
\name{plot.pedELRT}
\alias{plot.pedELRT}
\title{Function for plotting a pedELRT object}
\usage{
\method{plot}{pedELRT}(x, inc.exact = FALSE, legend = FALSE, ...)
}
\arguments{
\item{x}{a pedELRT object}

\item{inc.exact}{a logical of whether to include the exact ELRT in the plot (default=FALSE)}

\item{legend}{a logical of whether to include a legend in the plot (default=FALSE)}

\item{...}{Additional arguments.}
}
\description{
This function plots two graphs side-by-side of the ELRT and the power of the tests specified in the computeELRT functions
}
\details{
Plotting function for pedELRT object.  It plots the approximate or exact ELRT from what is calculated in the computeELRTX and computeEXLERT functions.
}
\examples{
require(kinship2)
test1 <- data.frame(id  =c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14),
mom =c(0, 0, 0, 0, 2, 2, 4, 4, 6,  2,  0,  0, 12, 13),
dad =c(0, 0, 0, 0, 1, 1, 3, 3, 3,  7,  0,  0, 11, 10),
sex =c(1, 2, 1, 2, 1, 2, 1, 2, 1, 1, 1, 2, 2, 2))
tped <- with(test1, pedigree(id, dad, mom, sex))
x <- computeRSPs(ped=tped)
elrtess <- computeAELRTESS(pedRSP=x,truehsq=0.3,null=seq(0.1,0.8,0.1))
plot(elrtess,inc.exact=TRUE,legend=TRUE)
}
\author{
Jesse D. Raffa
}
\seealso{
\code{\link[pedRSP]{computeAELRTA}} \code{\link[pedRSP]{computeAELRTB}} \code{\link[pedRSP]{computeAELRTC}} \code{\link[pedRSP]{computeAELRTESS}} \code{\link[pedRSP]{computeEXELRT}}
}

